requirements.txt
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.7.22
chardet==5.2.0
CheKnife==0.0.8
configparser==6.0.0
Django==4.2.5
DjTbot==0.0.7
exceptiongroup==1.1.3
future==0.18.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
netaddr==0.8.0
pycryptodome==3.18.0
python-dotenv==1.0.0
python-telegram-bot==13.15
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.1
typing_extensions==4.7.1
tzlocal==5.0.1
